split helper functions part 7
